+++ packages/embed-wallet/src/types.ts
@@ -23,6 +23,7 @@ export type ContextApp = {
   name?: string;
   logoUrl?: string;
   whiteLabel?: any; // TODO add proper type
+  skipOnboardingHelloPage?: boolean;
 };
 
 export type ContextWhiteLabel = any; // TODO: Figure out the type
@@ -83,6 +84,7 @@ export type WalletInitOptions = WalletOptions & {
   context?: PartialContext;
   popupMode?: 'popup' | 'modal';
   connectOptions?: Partial<WalletConnectOptions>;
+  skipOnboardingHelloPage?: boolean;
 };
 
 export type WalletTransferOptions = {