+++ src/routes/AuthorizationRouter.tsx
@@ -5,12 +5,13 @@ import { useAppContextStore } from '~/hooks';
 
 export const AuthorizationRouter = () => {
   const store = useAppContextStore();
+  const skipOnboardingHelloPage = Boolean(store.app?.skipOnboardingHelloPage);
 
   const isGame = ['metaverse-dash-run', 'candy-jam', 'cere-game-portal'].includes(store.app?.appId as string); // TODO remove after promo
   return (
     <Routes>
       <Route element={<Authorize />}>
-        <Route index element={isGame ? <LoginRoute /> : <IntroRoute />} />
+        <Route index element={isGame || skipOnboardingHelloPage ? <LoginRoute /> : <IntroRoute />} />
         <Route path="close" element={<AuthorizeClose />} />
         <Route path="redirect" element={<AuthorizeRedirect />} />
         <Route path="intro" element={<IntroRoute />} />